+++ src/adm_engine/adm_helper.cpp
