+++ src/adm_engine/audio_object_renderer.hpp
