+++ src/adm_engine/renderer.cpp
